+++ api/lib/application/certification-centers/certification-center-controller.js
@@ -2,6 +2,7 @@ import lodash from 'lodash';
 
 import { usecases as certificationConfigurationUsecases } from '../../../src/certification/configuration/domain/usecases/index.js';
 import * as divisionSerializer from '../../../src/prescription/campaign/infrastructure/serializers/jsonapi/division-serializer.js';
+import { usecases as teamUsecases } from '../../../src/team/domain/usecases/index.js';
 import { usecases } from '../../domain/usecases/index.js';
 import { DomainTransaction } from '../../infrastructure/DomainTransaction.js';
 import * as certificationCenterForAdminSerializer from '../../infrastructure/serializers/jsonapi/certification-center-for-admin-serializer.js';
@@ -102,7 +103,7 @@ const findCertificationCenterMembershipsByCertificationCenter = async function (
   dependencies = { certificationCenterMembershipSerializer },
 ) {
   const certificationCenterId = request.params.certificationCenterId;
-  const certificationCenterMemberships = await usecases.findCertificationCenterMembershipsByCertificationCenter({
+  const certificationCenterMemberships = await teamUsecases.findCertificationCenterMembershipsByCertificationCenter({
     certificationCenterId,
   });
 
@@ -115,7 +116,7 @@ const findCertificationCenterMemberships = async function (
   dependencies = { certificationCenterMembershipSerializer },
 ) {
   const certificationCenterId = request.params.certificationCenterId;
-  const certificationCenterMemberships = await usecases.findCertificationCenterMembershipsByCertificationCenter({
+  const certificationCenterMemberships = await teamUsecases.findCertificationCenterMembershipsByCertificationCenter({
     certificationCenterId,
   });
